+++ app/src/main/java/woowacourse/movie/service/Permission.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie
+package woowacourse.movie.service
 
 import android.Manifest
 import android.content.pm.PackageManager
